Added user_name global helper, consolidated helpers into global_helpers
