+++ lib/digital_public_works_web.ex
@@ -47,8 +47,7 @@ defmodule DigitalPublicWorksWeb do
 
       alias PhoenixBootstrapForm, as: PBF
       import DigitalPublicWorksWeb.Permission
-      import DigitalPublicWorksWeb.FormHelpers
-      import DigitalPublicWorksWeb.HTMLHelpers
+      import DigitalPublicWorksWeb.GlobalHelpers
     end
   end
