+++ lib/digital_public_works_web/templates/project/_project.html.eex
@@ -12,7 +12,7 @@
     <h5 class="card-title">
       <%= @project.title %>
       <small>
-        by <%= @project.user.display_name || "Project Member" %>
+        by <%= user_name(@project) %>
       </small>
     </h5>
     <div class="card-text mb-2"><%= @project.body %></div>
